MemoriesAPI interface created

+++ app/src/main/java/hu/bme/aut/moblab/memories/model/dto/ImageDTO.kt
@@ -4,7 +4,7 @@ import com.squareup.moshi.JsonClass
 
 @JsonClass(generateAdapter = true)
 data class ImageDTO (
-    val imageId: Int?,
-    val memoryId: Int?,
+    val imageId: Long?,
+    val memoryId: Long?,
     val url: String?
 )
